requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
celery
redis
python-docx
aamarpay
